Create a predeterminable address for EmoteRepository
The EmoteRepository contract is a common good smart contract that will
be used to store all the emotes. We want to be able to predict the
address of this contract so that we can use it on every network the
repository is deployed to. To do this, we will use a proxy contract that
will always deploy the same contract to the same address.